backend/requirements.txt
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.7
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.3.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
Werkzeug==3.0.3
WTForms==3.1.2
